test: harden RI A/B evidence chain for blind scoring
